Move Kinepolis Client to separate subFolder
