+++ src/Kinepolis/Client/KinepolisClient.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Kinepolis;
+namespace CultuurNet\UDB3\Kinepolis\Client;
 
 use GuzzleHttp\Psr7\UploadedFile;
